IS-1746: Move Avvent and Ikke Aktuell buttons
